First commit of Spectroperf (after history rewrite).
Spectroperf is a workload generator and benchmark intended to
allow measurement of mixed workloads, representing something
realistic, without coordinated omission.  See the README for a
more complete intro.

Note:

After a series of messy commits in the early days, this repository
started to take shape.  For initial publishing, we'll just pretend
it was done well at the outset.

The epitaph for the code that died to get here is in this commit
message (and on my machine, until git-gc).

Earlier commits:

snapshot

WIP: Updated to load and run from a workload definition.

Idea here is to be able to define workloads and run them from a
common main driver.

For this step, move the one time setup to a function and
create a run function for the workload.

Later, come back and grab the function names, possibly.  Or maybe
main is just a dumb driver.

WIP: Loads documents, demonstrates workload with CLI output.

Workloads are now in a separate file.  I'm not sure if this is how
it'll end up long term, but the thought is that workloads define some
functions for setup and run.

This now has the matrix working, albiet it's internal to the workload.

WIP: Random calling of workload operations from runners is good enough.

Workload implemented in all but search.

This has a main driver, workload implemented for KV and query operations
leaving out search for related profiles, which I think of as a proxy
for vector search but might just use with inverted indexes intiially.

Next steps are to add search and the index creation, then implement
the functional equivalent with data API. . Then some cleanup in
parameters, etc.

Fixed signal handling, added context for IO.

Move to modern file reading for golang.